requirements.txt
click==7.1.1
Flask==1.1.2
grpc-tools==0.0.1
grpcio==1.28.1
grpcio-tools==1.28.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.11.3
six==1.14.0
Werkzeug==1.0.1
